pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "walltext"
version = "0.1.0"
description = "Render centered text to a screen-sized PNG and set it as the Windows wallpaper."
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["Pillow>=10.0.0"]

[project.scripts]
walltext = "walltext.cli:main"

[tool.setuptools]
packages = ["walltext"]
include-package-data = true
